[parse::animated] Refactor single sections file into one file per section
